% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.2018}
\alias{dat.2018}
\title{2018 990 Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 64139 rows and 24 columns.
}
\usage{
dat.2018
}
\description{
All data used to generate \code{\link{model6}}.
Contains raw 990 data, features, and factor scores.
}
\keyword{datasets}
