% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features2}
\alias{features2}
\title{Features for Training Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 64139 rows and 12 columns.
}
\usage{
features2
}
\description{
Training data used to generate factor model \code{\link{model6}}.
See Step 5 of \code{vignette("making-gov-scores", package = "governance")} for detailed description of each of these features.
}
\keyword{datasets}
