% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model6}
\alias{model6}
\title{Factor Model}
\format{
An object of class \code{psych} (inherits from \code{fa}) of length 51.
}
\usage{
model6
}
\description{
Factor model object from \code{\link{features2}} data set using polychoric
correlations from \code{\link{rho2}}. This is the output of the \code{psych::fa} function.
This model has 12 features and 6 factors. See \code{vignette("making-gov-scores", package = "governance")}
for a detailed description of how and why this model was chosen.

Most notable internal objects include
\describe{
\item{loadings}{Loadings from the factor analysis using polychoric correlations of \code{features2}.}
\item{scores}{Scores using the PEARSON correlations. Note, we do not use the
scores generated from the \code{psych::fa} function as they always use the Pearson
correlation, even if the factor loadings were caclulated using the polychoric correlations.}
}
}
\keyword{datasets}
