% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores}
\alias{scores}
\title{Training Factor Scores}
\format{
An object of class \code{list} of length 5.
}
\usage{
scores
}
\description{
The factor scores from factor model \code{\link{model6}} of the  \code{\link{features2}} training data.
These are NOT the scores from \code{model6}, but are rather the factor scores calculated
using the \code{psych::facotor.scores} function using polychoric correlations
\code{\link{rho2}} and \code{method == "Thurstone"}. See step 6 of
\code{vignette("making-gov-scores", package = "governance")} for detailed
information on how this was generated.
}
\keyword{datasets}
