% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_example}
\alias{dat_example}
\title{Example Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 198729 rows and 33 columns.
}
\usage{
dat_example
}
\description{
Raw 990 data from 2013 to be used as an example for how to make features and run the factor model.
See \code{data-raw/01-get-example-data.R} and \code{vignette("download-data", package = "governance")} for details on how this data was downloaded.
Data is from NCCS 990 data download website \href{https://nccs.urban.org/nccs/catalogs/catalog-efile.html}{here}.
}
\keyword{datasets}
