% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-scores.R
\name{get_scores}
\alias{get_scores}
\title{Calculate Governance Scores}
\usage{
get_scores(feature.matrix)
}
\arguments{
\item{feature.matrix}{A matrix containing responses to questions for organizations.
Output from \code{get_features} function.
Rows represent each organization, and columns represent responses to the following questions:
"P12_LINE_1", "P4_LINE_12", "P4_LINE_28", "P4_LINE_29_30", "P6_LINE_1", "P6_LINE_11A",
"P6_LINE_15A", "P6_LINE_18", "P6_LINE_2", "P6_LINE_3", "P6_LINE_8A", "P6_LINE_12_13_14".
feature.matrix if a row has an NA value, the entire row will be removed in the output.}
}
\value{
A data frame with the original input data and appended 6 factor scores along with a total score.
}
\description{
This function takes a data.frame of features appends 6 factor scores
along with a total score.
}
\details{
This function generates factor scores for observations in the input \code{feature.matrix} from pre-loaded
factor model ( in \code{factor-objects.Rdata})
}
\references{
Factor objects are loaded from "governance/pkg-funcs/factor-objects.Rdata".
}
\seealso{
\code{\link{get_features}} for formatting \code{feature.matrix}.
}
